requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
pandas
requests
pytest